Refactor client proxy marshalling logic in wildfly-clustering-ejb-client.
