+++ ee-feature-pack/common/src/main/resources/modules/system/layers/base/org/wildfly/clustering/ejb/client/main/module.xml
@@ -35,6 +35,7 @@
     <dependencies>
         <module name="org.infinispan.protostream"/>
         <module name="org.jboss.ejb-client"/>
+        <module name="org.jboss.marshalling"/>
         <module name="org.wildfly.clustering.marshalling.api"/>
         <module name="org.wildfly.clustering.marshalling.spi"/>
         <module name="org.wildfly.clustering.marshalling.protostream"/>
